Refactor: 개행문자 추가
코드 파일관 일관성 유지
