+++ src/main/java/annotation/DataModel.java
@@ -8,5 +8,7 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
 public @interface DataModel {
+
     String path();
+
 }
