+++ src/main/java/db/Database.java
@@ -1,13 +1,13 @@
 package db;
 
 import com.google.common.collect.Maps;
-
 import model.User;
 
 import java.util.Collection;
 import java.util.Map;
 
 public class Database {
+
     private static Map<String, User> users = Maps.newHashMap();
 
     public static void clear() {
@@ -25,4 +25,5 @@ public class Database {
     public static Collection<User> findAll() {
         return users.values();
     }
+
 }
