+++ src/main/java/utils/StringUtils.java
@@ -7,4 +7,5 @@ public abstract class StringUtils {
     public static String appendNewLine(final String... messages) {
         return String.join(NEW_LINE, messages);
     }
+
 }
