+++ src/main/java/webserver/WebServer.java
@@ -9,6 +9,7 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
 public class WebServer {
+
     private static final Logger logger = LoggerFactory.getLogger(WebServer.class);
     private static final int DEFAULT_PORT = 8080;
     private static final int THREAD_NUMBER = 16;
@@ -34,4 +35,5 @@ public class WebServer {
             }
         }
     }
+
 }
